+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add JSP icon [Issue #32](https://github.com/BeardedBear/bearded-icons/issues/32)
 - Add MOV icon [Issue #33](https://github.com/BeardedBear/bearded-icons/issues/33)
 - Add MP4 icon [Issue #34](https://github.com/BeardedBear/bearded-icons/issues/34)
+- Add Rescript & BSconfig icons [Issue #39](https://github.com/BeardedBear/bearded-icons/issues/39)
 
 ## 0.14.0 - 2021-07-24